do not create directories recursively
`DirBuilder` ignores if the directory already exists. This is a bit
of an isuse security-wise. An adversary could try to create the
directory first, giving her full access to our temporary directory.

Simply disabling the recursive mode appears to be the simplest and
most secure solution. Another possibilities might be to use a more
secury RNG (slower!) or implementing directory creation ourselves,
failing if the directory already exists.