improve security and fix a few minor issues (#5)
* update uuid crate to 0.4.0

* do not create directories recursively

`DirBuilder` ignores if the directory already exists. This is a bit
of an isuse security-wise. An adversary could try to create the
directory first, giving her full access to our temporary directory.

Simply disabling the recursive mode appears to be the simplest and
most secure solution. Another possibilities might be to use a more
secury RNG (slower!) or implementing directory creation ourselves,
failing if the directory already exists.

* don't panic after file or directory creation failed

If creation of a file or directory fails, the destructor panics
because the file doesn't exist.

Example:

// Temp instance is created here
let temp = Temp {
    path: create_path_in(directory.to_path_buf()),
    _type: TempType::Dir,
    _released: false,
};

// `Temp` is dropped here if creation the directory fails and the destructor
// will fail to remove the directory since its creation failed.
try!(temp.create_dir());

* reformat using fmt